[efi] Enable NET_PROTO_IPV6 by default
IPv6 PXE was included in the UEFI specification over eight years ago,
specifically in version 2.3 (Errata D).

http://www.uefi.org/sites/default/files/resources/UEFI_Spec_2_3_D.pdf

When iPXE is being chainloaded from a UEFI firmware performing a PXE boot in an
IPv6 network, it is essential that iPXE supports IPv6 as well.

I understand that the reason for NET_PROTO_IPV6 being disabled by default (in
src/config/general.h) is that it would cause certain space-constrained build
targets to become too large. However, this should not be an issue for EFI
builds.

It is also worth noting that RFC 6540 makes a clear recommendation that IPv6
support should not be considered optional.

https://tools.ietf.org/html/rfc6540

Signed-off-by: Tore Anderson <tore@fud.no>