[serial] Allow serial port configuration at runtime
Add configuration variables for the serial port and its options,
allowing the configuration of a serial port at runtime via either "set",
or DHCP options.

The new "serial-port" option accepts values of "1" through "4" for COM1
to COM4 respectively. For the serial options, rather than adding a
configuration variable for each of them (baud, parity etc.) and
complicate both the code and the configuration unncessarily, add a
single, generic "serial-options" setting that accepts string values such
as "115200" or "38400n81". It also accepts the special value "preserve"
to preserve the previous configuration of the serial port.

Note that this commit does _not_ enable CONSOLE_SERIAL by default,
although that could potentially follow in a subsequent commit, given
this would be now a no-op unless a port is explicitly configured.

Signed-off-by: Faidon Liambotis <paravoid@debian.org>