[libc] Fix memcmp() to return proper values
Fix memcmp() to return proper standard positive/negative
values for unequal comparisons. Current implementation is backwards
(i.e. the functions are returning negative when should be positive and
 vice-versa).

Currently all consumers of these functions only check the return value
for ==0 or !=0 and so we can safely change the implementation without
breaking things. However, there is one call that checks the polarity of
this function, and that is prf_sha1 for wireless WPA 4-way
handshake. Due to the incorrect memcmp polarity, the WPA handshake
creates an incorrect PTK, and the handshake would fail after step 2.
Undoubtedly, the AP noticed the supplicant failed the mic check. This
commit fixes that issue.

Similar to commit 3946aa9bac1b4a784924fab1693bfebfdd357438

Signed-off-by: Michael Bazzinotti <bazz@bazz1.com>