image/multiboot2: pass kernel size and entry point to LZ in lz_set() call
LZ has to know the location and size of kernel so it can be measured. Assume
that only the first loadable section of ELF file is to be measured. mb2_bib
structure is not yet filled at the point of calling lz_set(), so it was used
to pass that information to keep the API simple.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>