[deflate] Add window buffer for GZIP decompression
GZIP decompression requires calculating the checksums, which, however,
needs all the uncompressed data. Repeated byte strings in the
uncompressed stream cannot be retrieved when there is no enough length
for the output buffer. This patch introduced a window buffer in the size
of GZIP_WSIZE, i.e. 32K, to hold the uncompressed data in a sliding
window manner.

Signed-off-by: Miao Wang <shankerwangmiao@gmail.com>