[image] Add support for GZIP compressed images
This patch introduced a new type of image entitled gzip, providing
support for gzip compressed boot images. After selected and exec-ed, the
image will be uncompressed into a new one with the same name and
cmdline, which immediately relpaces the original one and get exec-ed.

This patch is based on Petr Borsodi's work in pr #97

Acked-by: Petr Borsodi <petr.borsodi@gmail.com>
Signed-off-by: Miao Wang <shankerwangmiao@gmail.com>