[efi] Restore the TPL to the original one
iPXE added a workaround to raise TPL to TPL_CALLBACK always, but it
would cause UEFI firmware exit boot service callback with the same TPL
level cannot be executed. Worse, some UEFI modules set a pointer at exit
boot service callback and use later, then will encounter exception /
hang for invalid pointer which is not set to a correct one since the
exit boot service callback is not executed.

The solution is to restore the TPL to the original one but not always
set it to TPL_CALLBACK.