[net] Fallback to cross-signed cert if self-signed
Some TLS implementations will include the full cert chain (including the
root) in the initial handshake. In that situation, we would try to
validate the cert chain and then get caught up on that self-signed cert.
If that root cert was not provided, then we would have checked ipxe's
cross-signed roots to close the gap.

This change expands that cross-signed cert-downloading behavior to
include situations where we encounter a self-signed cert.

Closes #147

Signed-off-by: Josh McSavaney <me@mcsau.cc>